Move findThemeFilterTerm
